+++ DNSRdataTypeMX.java
@@ -26,7 +26,7 @@ public class DNSRdataTypeMX extends DNSRdata
 
     @Override
     /** implement the abstract method. */
-    public int getDataLength()
+    public short getDataLength()
     {
         //for research project, no need for implementation here.
         return 0;
