+++ DNSRdataTypeSOA.java
@@ -36,7 +36,7 @@ public class DNSRdataTypeSOA extends DNSRdata {
 
     @Override
     /** implement the abstract method. */
-    public int getDataLength()
+    public short getDataLength()
     {
         //for research project, no need for implementation here.
         return 0;
